requirement.txt
Django==4.0.6
django-extensions==3.2.0
ipython
django-admin-rangefilter
django-silk
django-import-export
django-crispy-forms
crispy-bootstrap5
django-filter
celery
djangorestframework
djangorestframework-simplejwt
drf-yasg

#dev
django-debug-toolbar

#flake
flake8==4.0.1
flake8-debugger==4.1.2
flake8-print==5.0.0
flake8-builtins==1.5.3

# pytest
pytest
pytest-django
pytest-cov
pytest-mock